conf: use mongo_collection and mongo_client instead of strings
